requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg>=0.28.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
fastapi>=0.103.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
paramiko>=3.3.1
pluggy==1.5.0
psycopg2-binary>=2.9.7
pycparser==2.22
pydantic>=2.3.0
pydantic_core==2.27.2
PyNaCl==1.5.0
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv>=1.0.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy>=2.0.20
sshtunnel==0.4.0
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
alembic>=1.12.0
email-validator>=2.0.0
python-multipart>=0.0.6
markdown>=3.4.4
jinja2>=3.1.2
uuid>=1.30
